Add a new method with better performance to traverse the text.
This new method checks, for each token, the next tokens defined by a threshold variable. The previous method is only commented for comparison.
An inflector class, with rules in Portuguese and English, is also added in order to singularize the tokens before querying the database, as concepts are usually in singular and the query performs exact matching.
In order to avoid GC overhead during the method testing, java memory was extended to 1024Mb.