Change recognized concepts to custom tag
The tag is changed to "x-health-translator" so the existent CSS rules of the target website don't override this element.
Fix a bug when there wasn't enough text on the webpage and it would get stuck in infinite loop. In that case, break the loop after a threshold of retries and assume english as default language.